In-place operations

In-place operations like i += 1, i -= 1 is not support in R. These functions implement these operations in R.

inc(e1, e2 = 1)

mult(e1, e2 = 2)

divi(e1, e2 = 2)

Arguments

e1
object, most likely a numeric object

e2
the operation value, the value to add, subtract, multiply, divide of.

Value

No return, directly assign the value back to e1

Details

inc(i) is the same as i <- i + 1. inc(i, -1) is the same as i <- i - 1. mult(i) is the same as i <- i * 2. divi(i) is the same as i <- i / 2.

See also

Examples

i <- 0
inc(i) # add 1
i
#> [1] 1
inc(i) # add 1
i
#> [1] 2
inc(i, -1) # minus 1
i
#> [1] 1
inc(i, -1) # minus 1
i
#> [1] 0
x <- 1
mult(x) # times 2
x
#> [1] 2
mult(x) # times 2
x
#> [1] 4
divi(x) # divide 2
x
#> [1] 2
divi(x) # divide 2
x
#> [1] 1